store the display plist data in raop_t as uint16_t and uint8_t types.

+++ lib/raop_handlers.h
@@ -134,13 +134,13 @@ raop_handler_info(raop_conn_t *conn,
     plist_t displays_0_uuid_node = plist_new_string("e0ff8a27-6738-3d56-8a16-cc53aacee925");
     plist_t displays_0_width_physical_node = plist_new_uint(0);
     plist_t displays_0_height_physical_node = plist_new_uint(0);
-    plist_t displays_0_width_node = plist_new_uint((uint16_t) conn->raop->display_width);
-    plist_t displays_0_height_node = plist_new_uint((uint16_t) conn->raop->display_height);
-    plist_t displays_0_width_pixels_node = plist_new_uint((uint16_t) conn->raop->display_width);
-    plist_t displays_0_height_pixels_node = plist_new_uint((uint16_t) conn->raop->display_height);
+    plist_t displays_0_width_node = plist_new_uint(conn->raop->display_width);
+    plist_t displays_0_height_node = plist_new_uint(conn->raop->display_height);
+    plist_t displays_0_width_pixels_node = plist_new_uint(conn->raop->display_width);
+    plist_t displays_0_height_pixels_node = plist_new_uint(conn->raop->display_height);
     plist_t displays_0_rotation_node = plist_new_bool(0);
-    plist_t displays_0_refresh_rate_node = plist_new_uint((uint8_t) conn->raop->display_refresh_rate);
-    plist_t displays_0_max_fps_node = plist_new_uint((uint8_t) conn->raop->display_max_fps);
+    plist_t displays_0_refresh_rate_node = plist_new_uint(conn->raop->display_refresh_rate);
+    plist_t displays_0_max_fps_node = plist_new_uint(conn->raop->display_max_fps);
     plist_t displays_0_overscanned_node = plist_new_bool(1);
     plist_t displays_0_features = plist_new_uint(14);
